change the page flow for failed passwords on the codebooks
Custom 401 ErrorDocument in Apache was not able to load the captcha
image. It has something to do with Apache not doing a full page load for
error docs. Just added a separate error doc that links to the
registration page, so users will not get stuck in a clunky workflow.